throttle: Move throttle code to its own .hh and .cc files
The move is very smoth, no changes in code needed but the definition of
throttle::capacity_t. It used to be the same as
fair_queue_entry::capacity_t, since we use the same notion of capacity
for both -- fair queuing and capacity throttling.
